Revert "Get end points"
