+++ domain/build.gradle
@@ -1,12 +1,15 @@
 
 dependencies {
     annotationProcessor 'org.projectlombok:lombok:1.18.12'
+
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.10.3'
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.10.3'
-    compile 'org.springframework.boot:spring-boot-starter-web:2.0.1.RELEASE'
-    compile 'org.springframework.boot:spring-boot-starter-data-jpa:2.0.1.RELEASE'
-   runtime("com.h2database:h2:1.3.176")
+
+
+
 }
+
+
 test {
     useJUnitPlatform()
 }
