+++ frontend/src/main/java/starter/project/frontend/Application.java
@@ -1,8 +1,7 @@
-package olumofe.raphael.frontend;
+package starter.project.frontend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 public class Application {
