+++ frontend/src/main/java/starter/project/frontend/Controller.java
@@ -1,10 +1,10 @@
-package olumofe.raphael.frontend;
+package starter.project.frontend;
 import org.springframework.web.bind.annotation.*;
+import starterProject.domain.*;
 import java.util.ArrayList;
 import java.util.List;
 @org.springframework.stereotype.Controller
-@RestController
-public class StatusController
+public class Controller
 {
     @GetMapping("/status")
     @ResponseBody
